Add a way to specify custom id types
NodeId, LinkId and PinId are just glorified pointers. In some cases
32/64 bits of data may not be enough to identify an object, or the
objects may lack a stable pointer/id value.
This commit adds a way to specify custom data types for them:
similarly to ImGui, a IMGUI_NODE_EDITOR_USER_CONFIG macro can be set
before including imgui_node_editor.h. If defined, its value will be
used as a filename to be included at the top of imgui_node_editor.h.

In it, the macros IMGUI_NODE_EDITOR_CUSTOM_NODEID,
IMGUI_NODE_EDITOR_CUSTOM_PINID and IMGUI_NODE_EDITOR_CUSTOM_LINKID
may be defined as the names of the types to be used for them.
If one or more of them are not defined the old types will be used,
old code does not need to be changed.

The types used must be copy contructible, assignable and must provide
a few more functions, as in this example:
struct CustomId
{
    CustomId(const CustomId &id);
    CustomId &operator=(const CustomId &id);;

    bool operator<(const CustomId &id) const;
    bool operator==(const CustomId &id) const;;

    std::string AsString() const;
    static CustomId FromString(std::string_view str);
    bool IsValid() const;

    static const CustomId Invalid;
};

A new example showcasing the functionality is added.